requirements.txt
blinker==1.8.2
Bootstrap-Flask==2.4.0
click==8.1.7
colorama==0.4.6
dominate==2.9.1
Flask==3.0.3
flask-bootstrap5==0.1.dev1
Flask-CKEditor==1.0.0
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
greenlet==3.0.3
itsdangerous==2.2.0
Jinja2==3.1.4
login==0.0.6
MarkupSafe==2.1.5
numpy==2.1.0
pandas==2.2.2
python-dateutil==2.9.0.post0
pytz==2024.1
six==1.16.0
SQLAlchemy==2.0.32
typing_extensions==4.12.2
tzdata==2024.1
visitor==0.1.3
Werkzeug==3.0.4
WTForms==3.1.2
